lots of fixes lots of visual upgrades terminal has per character line dynamic line break added log to display info and logging messages fixed network not scaling correctly added more colors

+++ src/main/java/gk646/jnet/neuralnetwork/Network.java
@@ -23,7 +23,7 @@ public final class Network {
     byte learnRate;
     byte outputLayerSize;
     float[][][] weightMatrix;
-    public short[] layerInfo;
+    public int[] layerInfo;
     //OBJECTS
     NetworkUtils netUtils = new NetworkUtils();
     Layer[] layers;
